Add .gitignore and .goreleaser.yaml files
This commit adds two new files, .gitignore and .goreleaser.yaml, to the repository. The .gitignore file is used to specify files and directories that should be ignored by Git. The .goreleaser.yaml file is an example configuration file for the Goreleaser tool, which is used for building and releasing Go applications.